Single Run command
The change modifies so that we have a single run command, rather than multiple.
Although it means that we now need to ignore the cases, it makes it simpler to
use as you don't need to decide which function to use based on which output
you want - you get all of them.

In the future we might add something like exec, where we know we don't expect
any output, and maybe this function will be renamed to out.